+++ packages/components/atoms/f-link/CHANGELOG.md
@@ -6,11 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.2.0
 ------------------------------
-*May 19, 2021*
+*May 20, 2021*
 
 ### Added
-- Props `isExternalLink` and `opensInNewLocation` to set `target`, `rel`, and `arialLabel`.
-- Props `isBold`, `isFullWidth` and `hadTextDecoration` to toggle link style.
+- Props `isExternalLink` and `opensInNewLocation` to set `target`, `rel`, and `ariaLabel`.
+- Props `isBold`, `isFullWidth` and `hasTextDecoration` to toggle link style.
 - Tests to cover changes.